pyproject.toml
[project]
name = "lecture-backend"
version = "0.1.0"
description = ""
authors = [
    { name = "Mahmoud Ayoub", email = "mayoub@edtpartners.com" },
    { name = "Eduardo López", email = "elopezv@edtpartners.com" },
    { name = "Lucas Caron", email = "lcaron@edtpartners.com" },
]
license = { text = "MIT" }
requires-python = ">=3.10,<4"
dependencies = [
    "numpy (>=1.26.4)",
    "fastapi[all] (>=0.115.12,<0.116.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "boto3 (>=1.37.34,<2.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "scikit-learn (>=0.24.2)",
    "requests (>=2.32.3,<3.0.0)",
    "icecream (>=2.1.4,<3.0.0)",
    "openai (>=1.74.0,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "starlette (>=0.46.2,<0.47.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "cryptography (>=44.0.2,<45.0.0)",
    "sqlalchemy (==1.4.25)",
    "psycopg2-binary (>=2.9.6)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "langchain (>=0.3.23,<0.4.0)",
    "langchain-community (>=0.3.21,<0.4.0)",
    "pdfminer-six (>=20250327,<20250328)",
    "pymupdf (>=1.23.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "langchain-text-splitters (>=0.3.8,<0.4.0)",
    "yt-dlp (>=2025.3.31,<2026.0.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "ffmpeg-python (>=0.2.0,<0.3.0)",
    "opensearch-py (>=2.8.0,<3.0.0)",
    "json-repair (==0.*)",
    "python-docx (>=1.1.2,<2.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "python-magic (==0.4.27)",
    "filetype (==1.2.0)",
    "langcodes (>=3.5.0,<4.0.0)",
    "ebooklib (>=0.18,<0.19)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pypandoc (>=1.15,<2.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "markdown (>=3.8,<4.0)",
    "language-data (>=1.3.0,<2.0.0)",
    "google-auth (>=2.39.0,<3.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "asyncpg (>=0.29.0)",
    "pywebpush (>=1.14.0,<2.0.0)",
    "greenlet (>=3.2.3,<4.0.0)",
    "pylti1p3 (>=2.0.0,<3.0.0)",
    "strands-agents (==1.4.0)",
    "strands-agents-tools (==0.2.3)",
    "httpx (>=0.28.1)",
    "lxml (>=4.9.4)",
    "SimpleIDML (==1.2.0)",
    "selenium (>=4.15.0,<5.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "8.3.4"
pytest-asyncio = "0.25.3"
pytest-cov = "6.0.0"
